Refactor product media utilities for better colocation
